Reuse code instead of duplicating
